Create WeatherReportModel to hold data from REST response.
WeatherReportModel is the model for holding the data returned from the API call. getCityForecastByID now returns a WeatherReportModel object, which contains all the weather information. A new responseListener was created for it.

Also removed the discord.xml file because for some reason adding it to the .gitignore did not work.